+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- [note] feature: improve the `performance` of `/note` command. (Use async method to fetch the player head data.)
+- feature: new module `command_meta.when_online`, to provide the `/when-online <player> <cmd>` command. (Execute the command exactly once when the target player online.)
+- [core] feature: now will `log the console` when failed to execute commands `as console`. (Make the debug easier.)